@@ -15,46 +15,44 @@
         }
     },
     "require": {
-        "php": ">=7.2",
+        "php": ">=8.0",
         "ext-dom": "*",
         "ext-json": "*",
         "ext-libxml": "*",
-        "cmgmyr/messenger": "^2.20",
-        "dingo/api": "~3.0.0",
-        "fideloper/proxy": "^4.3",
-        "fruitcake/laravel-cors": "^1.0",
-        "guzzlehttp/guzzle": "^6.3",
-        "intervention/image": "^2.5",
-        "irazasyed/telegram-bot-sdk": "^3.1.0",
-        "laravel/framework": "^7.1",
+        "cmgmyr/messenger": "^2.16.2",
+        "dingo/api": "~3.0",
+        "fideloper/proxy": "^4.4",
+        "fruitcake/laravel-cors": "^2.0",
+        "guzzlehttp/guzzle": "^7.3.0",
+        "intervention/image": "^2.6",
+        "irazasyed/telegram-bot-sdk": "^v3.4.1",
+        "laravel/framework": "^8.47",
         "laravel/tinker": "~2.0",
-        "laravel/ui": "^2.0",
+        "laravel/ui": "^3.3",
         "maatwebsite/excel": "^3.1.19",
-        "medcenter24/lib-installer": "~2.0",
+        "medcenter24/lib-installer": "^2.0.1",
         "mpdf/mpdf": "~8.0.6",
-        "nwidart/laravel-modules": "7.1.0",
-        "spatie/laravel-medialibrary": "~8.2.1",
+        "nwidart/laravel-modules": "^8.2",
+        "spatie/laravel-medialibrary": "^9.6",
         "srmklive/flysystem-dropbox-v2": "1.0.7",
-        "tymon/jwt-auth": "1.0.0"
+        "tymon/jwt-auth": "^1.0.2"
     },
     "require-dev": {
-        "barryvdh/laravel-debugbar": "~3.3.2",
-        "barryvdh/laravel-ide-helper": "^2.7.0",
+        "barryvdh/laravel-debugbar": "^3.6",
+        "barryvdh/laravel-ide-helper": "^2.10",
         "filp/whoops": "^2.7.1",
-        "facade/ignition": "^2.0.2",
-        "fzaninotto/faker": "~1.9.1",
-        "mockery/mockery": "^1.3",
-        "nunomaduro/collision": "^4.1",
-        "phpunit/phpunit": "^8.5",
+        "facade/ignition": "^2.8.3",
+        "fzaninotto/faker": "^1.9",
+        "mockery/mockery": "^1.4",
+        "nunomaduro/collision": "^5.5",
+        "phpunit/phpunit": "^9.5",
         "roave/security-advisories": "dev-master"
     },
     "autoload": {
-        "classmap": [
-            "database/seeds",
-            "database/factories"
-        ],
         "psr-4": {
-            "medcenter24\\mcCore\\App\\": "app"
+            "medcenter24\\mcCore\\App\\": "app/",
+            "Database\\Factories\\": "database/factories/",
+            "Database\\Seeders\\": "database/seeds/"
         }
     },
     "autoload-dev": {
