@@ -15,6 +15,9 @@
  *
  * Copyright (c) 2019 (original work) MedCenter24.com;
  */
+declare(strict_types = 1);
+
+namespace Database\Seeders;
 
 use medcenter24\mcCore\App\Entity\FinanceCondition;
 use Illuminate\Database\Seeder;
@@ -24,6 +27,6 @@ class FinanceConditionTableSeeder extends Seeder
     public function run()
     {
         FinanceCondition::truncate();
-        factory(FinanceCondition::class, 3)->create();
+        FinanceCondition::factory()->count(3)->create();
     }
 }
