@@ -15,7 +15,11 @@
  *
  * Copyright (c) 2019 (original work) MedCenter24.com;
  */
+declare(strict_types = 1);
 
+namespace Database\Seeders;
+
+use Illuminate\Support\Facades\App;
 use medcenter24\mcCore\App\Entity\Form;
 use medcenter24\mcCore\App\Entity\FormReport;
 use Illuminate\Database\Seeder;
@@ -33,9 +37,9 @@ class FormReportsTableSeeder extends Seeder
             return;
         } elseif (!App::environment('production')) {
             FormReport::truncate();
-            factory(FormReport::class, 10)->create([
+            FormReport::factory()->count(10)->create([
                 'form_id' => function () {
-                    return factory(Form::class)->create([
+                    return Form::factory()->create([
                         'title' => 'Accident Report Form',
                         'description' => 'Form has been generated by ModelFactory for FormReport',
                         'template' => '<p>Form for :className</p><p>Generated by <b>:generatorName</b></p>',
