@@ -20,10 +20,10 @@ declare(strict_types = 1);
 
 namespace medcenter24\mcCore\Tests\Feature\Api\Director\Cases;
 
+use Database\Seeders\ScenariosTableSeeder;
 use medcenter24\mcCore\App\Services\Entity\CaseAccidentService;
 use medcenter24\mcCore\Tests\Feature\Api\DirectorTestTraitApi;
 use medcenter24\mcCore\Tests\TestCase;
-use ScenariosTableSeeder;
 
 class CaseStoryControllerTest extends TestCase
 {
