@@ -17,7 +17,7 @@
           window.Laravel = {csrfToken: '{{ csrf_token() }}'};
         </script>
 
-        <link href="{{ elixir('/css/app.css') }}" rel="stylesheet">
+        <link href="{{ mix('/css/app.css') }}" rel="stylesheet">
     </head>
     <body>
         <div class="flex-center position-ref full-height">
@@ -45,7 +45,7 @@
             </div>
         </div>
 
-        <script src="{{ elixir('/js/app.js') }}"></script>
+        <script src="{{ mix('/js/app.js') }}"></script>
 
     </body>
 </html>
